Restructure Biomass-Harvest
Restructured Biomass-Harvest so as not to require the use of an SDK or
that the DLLs be stored in the repository. Removed some unnecessary
files and added some files to download and put the necessary DLLs in the
correct place. Biomass-Harvest should now pull all the necessary DLLs
when building.